HTTPSourceConnectionの共通化できる部分を括り出した
HTTPSourceConnectionの共通化できる部分を他でも使うため括り出してSourceConnectionBase2とした。
そのうち2を外してSourceConnectionBaseを乗っ取る予定。
